move navbarScrollResponsive and menuOpenClose functions to their own files so that they can be imported to different pages
