+++ src/features/navbarMenuOpenClose.js
@@ -1,0 +1,12 @@
+const mainMenuContainer = document.getElementById("mainMenuContainer");
+const navbarButton = document.getElementById("navbarButton");
+
+export function menuOpenClose() {
+  if (mainMenuContainer.style.display === "flex") {
+    mainMenuContainer.style.display = "none";
+    navbarButton.classList.toggle("clicked");
+  } else {
+    mainMenuContainer.style.display = "flex";
+    navbarButton.classList.toggle("clicked");
+  }
+}
